Refactored course overview to make it more harmonious with other codes

+++ src/main/java/org/amadeus/charon/data/Course.java
@@ -7,7 +7,7 @@ public class Course {
 		String courseName;
 		String courseDesc;
 		
-		public Course(int i, String c, String n, String d){
+		public Course(long i, String c, String n, String d){
 			this.id = i;
 			this.courseCode = c;
 			this.courseName = n;
